+++ src/main.rs
@@ -3,11 +3,12 @@ use config::Config;
 use dirs::config_dir;
 use err::Result;
 use maker::{copy_dir, create_template, load_template};
+use pareg::Pareg;
 use std::{
     borrow::Cow,
-    env,
     fs::{read_dir, remove_dir_all},
     io::{stderr, stdin, stdout, IsTerminal, Write},
+    mem,
     path::{Path, PathBuf},
     process::ExitCode,
 };
@@ -36,15 +37,12 @@ fn main() -> ExitCode {
 }
 
 fn start() -> Result<()> {
-    let args: Vec<_> = env::args().collect();
-    let args = Args::parse(args.iter().skip(1).into())?;
-
-    args.check_unused();
+    let mut args = Args::parse(Pareg::args())?;
 
     // Do what the arguments specify
-    match args.get_action() {
+    match mem::take(&mut args.action) {
         Action::Create => create(args)?,
-        Action::Alias => alias(args)?,
+        Action::Alias(al) => alias(args, al)?,
         Action::Config => configure(args)?,
         Action::Load => load(args)?,
         Action::Remove => remove(args)?,
@@ -60,16 +58,14 @@ fn start() -> Result<()> {
 /// Creates new template with the name `name` from the directory `src` in the
 /// default template folder.
 fn create(args: Args) -> Result<()> {
-    let name = args.get_template()?;
-    let src = args.get_directory();
-
-    let tdir = get_template_dir(name)?;
+    let tdir = get_template_dir(&args.template)?;
 
     if Path::new(&tdir).exists() {
         if !prompt_yn(
             &format!(
-                "Template with the name '{name}' already \
-                exists.\nDo you want to overwrite it? [y/N]: "
+                "Template with the name '{}' already \
+                exists.\nDo you want to overwrite it? [y/N]: ",
+                args.template
             ),
             args.prompt_answer,
         )? {
@@ -78,17 +74,15 @@ fn create(args: Args) -> Result<()> {
         remove_dir_all(&tdir)?;
     }
 
-    create_template(src, tdir)
+    create_template(args.directory.as_ref(), tdir)
 }
 
 /// Creates new template with the name `name` from the directory `src` in the
 /// default template folder.
-fn alias(args: Args) -> Result<()> {
-    let template = args.get_template()?;
-    let alias_name = args.get_alias()?;
+fn alias(args: Args, alias_name: String) -> Result<()> {
     let mut conf = load_config()?;
 
-    if conf.aliases.contains_key(alias_name)
+    if conf.aliases.contains_key(&alias_name)
         && !prompt_yn(
             &format!(
                 "Alias with the name '{alias_name}' already exists.\n\
@@ -101,12 +95,8 @@ fn alias(args: Args) -> Result<()> {
     }
 
     let alias = Alias {
-        template: template.to_owned(),
-        vars: args
-            .vars
-            .into_iter()
-            .map(|(k, v)| (k.into_owned().into(), v.into_owned().into()))
-            .collect(),
+        template: args.template,
+        vars: args.vars,
     };
 
     conf.aliases.insert(alias_name.to_owned(), alias);
@@ -121,10 +111,9 @@ fn configure(args: Args) -> Result<()> {
         if let Some(k) = k.strip_prefix('-') {
             conf.vars.remove(k);
         } else if let Some(k) = k.strip_prefix('+') {
-            conf.vars.insert(k.to_owned().into(), v.into_owned().into());
+            conf.vars.insert(k.to_string().into(), v);
         } else {
-            conf.vars
-                .insert(k.into_owned().into(), v.into_owned().into());
+            conf.vars.insert(k, v);
         }
     }
 
@@ -134,19 +123,17 @@ fn configure(args: Args) -> Result<()> {
 /// Loads template with the name `src` to the directory `dest`. `vars` can
 /// add/override variables in the template config file.
 fn load(mut args: Args) -> Result<()> {
-    let mut name = args.get_template()?;
+    let mut name: Cow<str> = args.template.into();
     let conf = load_config()?;
 
-    if let Some(a) = conf.aliases.get(name) {
-        name = &a.template;
+    if let Some(a) = conf.aliases.get(name.as_ref()) {
+        name = a.template.as_str().into();
         for (k, v) in &a.vars {
             args.vars.entry(k.clone()).or_insert(v.clone());
         }
     }
 
-    let dest = args.get_directory();
-
-    let template = get_template_dir(name)?;
+    let template = get_template_dir(&name)?;
     if !template.exists() {
         return Err(Error::Msg(
             format!("There is no existing template '{name}'").into(),
@@ -154,11 +141,15 @@ fn load(mut args: Args) -> Result<()> {
     }
 
     // true if the directory exists and isn't empty
-    if read_dir(dest).ok().and_then(|mut d| d.next()).is_some()
+    if read_dir(args.directory.as_ref())
+        .ok()
+        .and_then(|mut d| d.next())
+        .is_some()
         && !prompt_yn(
             &format!(
-                "the directory {dest} is not empty.\n\
-            Do you want to load the template anyway? [y/N]: "
+                "the directory {} is not empty.\n\
+                Do you want to load the template anyway? [y/N]: ",
+                args.directory
             ),
             args.prompt_answer,
         )?
@@ -166,60 +157,61 @@ fn load(mut args: Args) -> Result<()> {
         return Ok(());
     }
 
-    let mut vars = args.vars;
     for (k, v) in conf.vars {
-        vars.entry(k).or_insert(v);
+        args.vars.entry(k).or_insert(v);
     }
 
-    load_template(template, dest, vars)
+    load_template(template, args.directory.as_ref(), args.vars)
 }
 
 /// Deletes template with the name `name`
 fn remove(args: Args) -> Result<()> {
     let mut conf = load_config()?;
-    if conf.aliases.remove(args.get_template()?).is_some() {
+    if conf.aliases.remove(&args.template).is_some() {
         save_config(&conf)
     } else {
-        remove_dir_all(get_template_dir(args.get_template()?)?)?;
+        remove_dir_all(get_template_dir(&args.template)?)?;
         Ok(())
     }
 }
 
 /// Copies the template with the name `name` to the directory `dest`.
 fn edit(args: Args) -> Result<()> {
-    let name = args.get_template()?;
-    let dest = args.get_directory();
-
-    let template = get_template_dir(name)?;
+    let template = get_template_dir(&args.template)?;
     if !template.exists() {
         let conf = load_config()?;
-        if let Some(n) = conf.aliases.get(name) {
+        if let Some(n) = conf.aliases.get(&args.template) {
             return Err(Error::Msg(
                 format!(
-                    "{name} is not a template, it is alias for the template \
+                    "{} is not a template, it is alias for the template \
                     {}",
-                    n.template
+                    args.template, n.template
                 )
                 .into(),
             ));
         }
         return Err(Error::Msg(
-            format!("There is no existing template '{name}'").into(),
+            format!("There is no existing template '{}'", args.template)
+                .into(),
         ));
     }
 
-    if read_dir(dest).ok().and_then(|mut d| d.next()).is_some()
+    if read_dir(args.directory.as_ref())
+        .ok()
+        .and_then(|mut d| d.next())
+        .is_some()
         && !prompt_yn(
             &format!(
-                "the directory {dest} is not empty.\n\
-            Do you want to load the template source anyway? [y/N]: "
+                "the directory {} is not empty.\n\
+                Do you want to load the template source anyway? [y/N]: ",
+                args.directory,
             ),
             args.prompt_answer,
         )?
     {
         return Ok(());
     }
-    copy_dir(template, dest)
+    copy_dir(template, args.directory.as_ref())
 }
 
 /// Prints all the template name to the stdout.